+++ lib/screens/splash_screen.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:videosdk_flutter_example/constants/colors.dart';
-import 'startup_screen.dart';
+import 'join_screen.dart';
 
 class SplashScreen extends StatefulWidget {
   const SplashScreen({Key? key}) : super(key: key);
@@ -19,7 +19,7 @@ class _SplashScreenState extends State<SplashScreen> {
       Navigator.pushReplacement(
         context,
         MaterialPageRoute(
-          builder: (context) => const StartupScreen(),
+          builder: (context) => const JoinScreen(),
         ),
       );
     });
