+++ lib/widgets/one_to_one_participant_view/participant_view_one_to_one.dart
@@ -1,15 +1,7 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
-import 'package:flutter/src/foundation/key.dart';
-import 'package:flutter/src/widgets/framework.dart';
 import 'package:flutter/widgets.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:videosdk/videosdk.dart';
 import 'package:videosdk_flutter_example/constants/colors.dart';
-import 'package:videosdk_flutter_example/utils/spacer.dart';
-import 'package:videosdk_flutter_example/widgets/one_to_one_participant_view/local_screen_share_view.dart';
-import 'package:videosdk_flutter_example/widgets/one_to_one_participant_view/participant_view.dart';
 import 'package:videosdk_flutter_example/widgets/one_to_one_participant_view/participant_view.dart';
 
 class ParticipantViewOneToOne extends StatefulWidget {
